# api/api.yml
openapi: "3.1.1"
info:
  title: Log-It API
  version: "1.0"
  description: API for developers to log and manage their application logs.
servers:
  - url: https://logit.amsky.xyz
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /auth/register:
    post:
      summary: Register for a dev key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
      responses:
        "201":
          description: Dev key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  publicKey:
                    type: string
                  token:
                    type: string
        "400":
          description: Request did not contain a password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  
  /auth/login:
    post:
      summary: Fetch a JWT token for auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                publicKey:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "400":
          description: Request did not contain password or public key
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "403":
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /api/pull:
    get:
      summary: Fetch logs
      description: |
        Fetch logs for the authenticated user or public logs with limited info.

        - With auth you get full filters and full data.
        - If no auth, provide `key` (public dev key) query parameter to fetch public logs with limited info.
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: query
          required: false
          description: Public dev key, required if no auth token is provided.
          schema:
            type: string
        - name: channel
          in: query
          required: false
          description: Filter logs by channel (only works with auth token).
          schema:
            type: string
        - name: ip
          in: query
          required: false
          description: Filter logs by IP address (only works with auth token).
          schema:
            type: string
        - name: logLevel
          in: query
          required: false
          description: Filter logs by log level (only works with auth token).
          schema:
            type: string
        - name: messageContains
          in: query
          required: false
          description: Filter logs containing this string in message (only works with auth token).
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Logs fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  publicKey:
                    type: string
                  page:
                    type: integer
                  logs:
                    type: array
                    items:
                      type: object
                      properties:
                        message:
                          type: string
                        channel:
                          type: string
                          description: |
                            Channel name. Will be "***" for public access.
                        ip:
                          type: string
                          description: |
                            IP address. Will be "***" for public access.
                        logLevel:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
                        logId:
                          type: string
                          description: |
                            Log ID. Will be "***" for public access.
        "401":
          description: Invalid or missing auth token when given
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "403":
          description: Invalid public key or invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/log:
    get:
      summary: Log a message
      parameters:
        - name: key
          description: Public dev key
          in: query
          required: true
          schema:
            type: string
        - name: channel
          in: query
          required: true
          schema:
            type: string
        - name: message
          in: query
          required: true
          schema:
            type: string
        - name: logLevel
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Log created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
        "400":
          description: Bad request, missing fields or invalid data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "403":
          description: Missing/invalid dev key
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/edit:
    put:
      summary: Edit a log's message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                logId:
                  type: string
                message:
                  type: string
      responses:
        "200":
          description: Log edited successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
        "400":
          description: Bad request, missing fields or invalid data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "403":
          description: Unauthorized access
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "404":
          description: Log not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/delete:
    delete:
      summary: Delete a log
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                logId:
                  type: string
      responses:
        "200":
          description: Log deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
        "400":
          description: Bad request, missing fields or invalid data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "403":
          description: Unauthorized access
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "404":
          description: Log not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/clear:
    delete:
      summary: Clear all logs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All logs cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  deletedCount:
                    type: integer
                    example: 50
        "403":
          description: Unauthorized access
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "404":
          description: No logs found to clear
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/export:
    get:
      summary: Export logs
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          required: false
          description: Format to export logs (default JSON)
          schema:
            type: string
            enum: [json, csv]
      responses:
        "200":
          description: Logs exported successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: 
                    {
                      "channel": "channel",
                      "ip": "1.1.1.1",
                      "logId": "abc",
                      "logLevel": "info",
                        "message": "message",
                        "timestamp": "2025-01-01T00:00:00.000000+00:00"
                      }
  /api/log/bulk:
    post:
      summary: Bulk log messages
      parameters:
        - name: key
          in: query
          required: true
          description: Public dev key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                properties:
                  channel:
                    type: string
                  message:
                    type: string
                  logLevel:
                    type: string
      responses:
        "200":
          description: Logs created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
        "400":
          description: Bad request, missing fields or invalid data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        "403":
          description: Unauthorized access
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /stats:
    get:
      summary: Fetch stats
      responses:
        "200":
          description: Successfully fetched stats
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalLogs:
                    type: integer
                    example: 1000
                  totalDevKeys:
                    type: integer
                    example: 10
        